initial commit for public release
In the 17th Chapter of St Luke it is written: “the Kingdom of God is
within man” - not one man nor a group of men, but in all men! In you!
You, the people have the power - the power to create machines. The power
to create happiness! You, the people, have the power to make this life
free and beautiful, to make this life a wonderful adventure.
                                         -- The Great Dictator